refactor: migrate GraphQLTest to use graphql() function pattern
  - Replace DocumentNode imports with inline graphql() function calls
  - Remove unused useGraphQL hook and queries.graphql file
  - Add "Test" prefix to query names to avoid conflicts with other components
  - Switch from getQueryString() to print() for query string extraction
